Factor out wasm types and parser declarations into their own files
First refactoring toward making the parser usable for things other than
the sexpr-to-wasm encoder. It moves the wasm types into their own
header, the parser types and declarations into their own header, and the
parser code into its own source file separate from main and the CLI
option parsing. Makes no attempt to actually change the interface yet.
